+++ Readme.md
@@ -1,6 +1,6 @@
 # Личный проект «Что посмотреть»2
 
-* Студент: [Константин Реент](https://up.htmlacademy.ru/univer-js3/3/user/2221625). 
+* Студент: [Константин Реент](https://up.htmlacademy.ru/univer-js3/3/user/2221625).
 * Наставник: `Владимир Ульянов`.
 
 ---
