+++ src/pages/add-review/add-review.test.tsx
@@ -51,6 +51,7 @@ describe('Component: AuthScreen', () => {
     await userEvent.clear(screen.getByTestId('textarea-test'));
     await userEvent.type(
       screen.getByTestId('textarea-test'),
+
       reviewText,
     );
     await userEvent.click(screen.getByTestId('test-star10'));
