+++ src/pages/movie-page/movie-page.tsx
@@ -2,7 +2,7 @@ import { Link } from 'react-router-dom';
 import { useNavigate } from 'react-router-dom';
 import Tabs from '../../components/tabs/tabs';
 import FilmList from '../../components/film-list/film-list';
-import MyListButton from '../../components/mylist-button/mylist-button';
+import MyListButton from '../../components/myList-button/myList-button';
 import { fetchCurrentFilm, fetchCurrentFilmRecomends, fetchCurrentFilmReviews } from '../../store/api-actions';
 import { useEffect, useCallback } from 'react';
 import { useParams } from 'react-router-dom';
